Move content endpoint functional tests to integration
The integration tests are more robust and test more things than the
functional tests here.
